+++ src/back/Plants.Application/Contracts/IPlantsService.cs
@@ -6,8 +6,8 @@ namespace Plants.Application.Contracts;
 public interface IPlantsService
 {
     Task<IEnumerable<PlantResultItem>> GetNotPosted();
-    Task<PreparedPostResultItem?> GetPrepared(int plantId);
-    Task<CreatePostResult> Post(int plantId, decimal price);
+    Task<PreparedPostResultItem?> GetPrepared(long plantId);
+    Task<CreatePostResult> Post(long plantId, decimal price);
     Task<AddPlantResult> Create(string Name, string Description,
         long[] Regions, long SoilId,
         long GroupId, DateTime Created,
@@ -16,5 +16,5 @@ public interface IPlantsService
     Task Edit(long PlantId, string Name, string Description,
             long[] Regions, long SoilId,
             long GroupId, long[] RemovedImages, byte[][] NewImages);
-    Task<PlantResultDto?> GetBy(int id);
+    Task<PlantResultDto?> GetBy(long id);
 }
